Fix mistakes
