+++ module_20/data/authorization.php
@@ -3,6 +3,6 @@
 namespace Module\Twenty;
 
 if (isset($_POST['login']) && isset($_POST['password'])) {
-    $user = new Authorization(Db::getConnection());
+    $user = new UserRepository(Db::getConnection());
     $user->validateUser();
 }
