+++ module_20/docker-compose.yaml
@@ -3,7 +3,7 @@ version: "3"
 services:
   db:
     image: mysql:8.0.33
-    platform: linux/x86_64
+    #platform: linux/x86_64
     command: --default-authentication-plugin=mysql_native_password
     volumes:
       - "db_app:/var/lib/mysql"
@@ -22,4 +22,4 @@ services:
       - 8080:8080
 
 volumes:
-    db_app:
+  db_app:
